requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
coverage==7.6.0
cryptography==42.0.8
decorator==5.1.1
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-spectacular==0.27.2
execnet==2.1.1
executing==2.0.1
Faker==26.0.0
filelock==3.15.4
identify==2.6.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
ipython==8.26.0
jedi==0.19.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Markdown==3.6
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
model-bakery==1.18.2
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.7.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-env==1.1.3
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rpds-py==0.19.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
stack-data==0.6.3
stripe==8.6.0
traitlets==5.14.3
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
whitenoise==6.7.0
